// src/exception_filter/errors.enum.ts
export enum ErrorCode {
    // Auth Validation
    MUST_BE_STRING = "MUST_BE_STRING",
    USERNAME_SHORT_LENGTH = "USERNAME_SHORT_LENGTH",
    USERNAME_LONG_LENGTH = "USERNAME_LONG_LENGTH",
    USERNAME_INVALID_FORMAT = "USERNAME_INVALID_FORMAT",
    PASSWORD_SHORT_LENGTH = "PASSWORD_SHORT_LENGTH",
    PASSWORD_LONG_LENGTH = "PASSWORD_LONG_LENGTH",
    FIRSTNAME_SHORT_LENGTH = "FIRSTNAME_SHORT_LENGTH",
    FIRSTNAME_LONG_LENGTH = "FIRSTNAME_LONG_LENGTH",
    LASTNAME_SHORT_LENGTH = "LASTNAME_SHORT_LENGTH",
    LASTNAME_LONG_LENGTH = "LASTNAME_LONG_LENGTH",

    // Auth Methods
    USERNAME_TAKEN = "USERNAME_TAKEN"
}